moved food files into components folder to to fix conflict
