+++ src/client/main.jsx
@@ -7,9 +7,9 @@ import App from "./App";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 // Layout/Components Imports
-import dessertList from "./views/layouts/dessertList";
-import entreeList from "./views/layouts/entreeList";
-import mealList from "./views/layouts/mealList" 
+import dessertList from "./components/views/dessertList";
+import entreeList from "./components/views/entreeList";
+import mealList from "./components/views/mealList" 
 
 // Routes
